FoundationSync: move sync code outside plugin.py
